requirements.txt
altair==5.5.0
anyio==3.7.1
asgiref==3.8.1
attrs==25.3.0
awscli==1.38.18
blinker==1.9.0
boto3==1.37.18
botocore==1.37.18
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
docutils==0.16
fastapi==0.68.2
filelock==3.18.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
huggingface-hub==0.29.3
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lancedb==0.21.1
MarkupSafe==3.0.2
mpmath==1.3.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pydantic==1.10.21
pydeck==0.9.1
pylance==0.24.1
PyMuPDF==1.25.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rpds-py==0.23.1
rsa==4.7.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
setuptools==77.0.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.14.2
streamlit==1.43.2
sympy==1.13.1
tantivy==0.22.2
tenacity==9.0.0
threadpoolctl==3.6.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.15.0
watchdog==6.0.0
